feat: action expandible component and graphical fixes in parent
