+++ src/ui-actions/mail-message-preview-actions.tsx
@@ -4,7 +4,6 @@
  * SPDX-License-Identifier: AGPL-3.0-only
  */
 import React, {
-	useContext,
 	useMemo,
 	FC,
 	ReactElement,
@@ -12,207 +11,81 @@ import React, {
 	useLayoutEffect,
 	useState,
 	useCallback,
-	useEffect
+	useContext
 } from 'react';
-import {
-	Row,
-	IconButton,
-	SnackbarManagerContext,
-	Tooltip,
-	useModal,
-	Dropdown,
-	ThemeContext
-} from '@zextras/carbonio-design-system';
-import { difference, includes, map, slice } from 'lodash';
-import { useTranslation } from 'react-i18next';
-import { useDispatch } from 'react-redux';
-import {
-	useIntegratedComponent,
-	useReplaceHistoryCallback,
-	useAppContext,
-	FOLDERS
-} from '@zextras/carbonio-shell-ui';
-
-import {
-	deleteMsg,
-	editAsNewMsg,
-	editDraft,
-	forwardMsg,
-	moveMsgToTrash,
-	redirectMsg,
-	replyAllMsg,
-	replyMsg,
-	sendDraft,
-	setMsgFlag,
-	setMsgAsSpam,
-	printMsg,
-	showOriginalMsg,
-	setMsgRead,
-	moveMessageToFolder,
-	deleteMessagePermanently
-} from './message-actions';
-import { MailMessage } from '../types/mail-message';
-import { useSelection } from '../hooks/useSelection';
+import { Row, IconButton, Tooltip, Dropdown, ThemeContext } from '@zextras/carbonio-design-system';
+import { difference, map, slice } from 'lodash';
+import { useVisibleActionsCount } from '../hooks/use-visible-actions-count';
 
 type MailMsgPreviewActionsType = {
-	folderId: string;
-	message: MailMessage;
-	timezone: string;
+	actions: Array<any>;
+	maxActions?: number;
+	maxWidth?: string;
+	mainAlignment?: string;
 };
 
-function useOverflowCount(containerRef: React.RefObject<HTMLInputElement>): [number, () => void] {
-	const [visibleActionsCount, setVisibleActionsCount] = useState<number>(0);
-	const theme = useContext(ThemeContext);
-	// eslint-disable-next-line @typescript-eslint/ban-ts-comment
-	// @ts-ignore
-	const iconSize = parseInt(theme.sizes.icon.large, 10);
-	const calculateVisibleActionsCount = useCallback(() => {
-		if (containerRef && containerRef.current && containerRef?.current?.clientWidth > 0) {
-			const value = Math.floor(containerRef.current.clientWidth / iconSize);
-			setVisibleActionsCount(value);
-		}
-	}, [containerRef, iconSize]);
-
-	useEffect(() => {
-		window.addEventListener('resize', calculateVisibleActionsCount);
-		return (): void => window.removeEventListener('resize', calculateVisibleActionsCount);
-	}, [calculateVisibleActionsCount]);
-
-	useEffect(() => {
-		window.addEventListener('transitionend', calculateVisibleActionsCount);
-		return (): void => window.removeEventListener('transitionend', calculateVisibleActionsCount);
-	}, [calculateVisibleActionsCount]);
-
-	return [visibleActionsCount, calculateVisibleActionsCount];
-}
+type ThemeContextProps = {
+	sizes: {
+		icon: {
+			large: string;
+		};
+	};
+};
 
 const MailMsgPreviewActions: FC<MailMsgPreviewActionsType> = ({
-	message,
-	folderId,
-	timezone
+	actions,
+	maxActions,
+	maxWidth = '120px',
+	mainAlignment = 'flex-end'
 }): ReactElement => {
-	const [t] = useTranslation();
-	const replaceHistory = useReplaceHistoryCallback();
-	const createSnackbar = useContext(SnackbarManagerContext);
-	const dispatch = useDispatch();
-	const createModal = useModal();
-	const ContactInput = useIntegratedComponent('contact-input');
-	const { setCount } = useAppContext();
-	const { deselectAll } = useSelection(folderId, setCount);
-	const systemFolders = useMemo(
-		() => [FOLDERS.INBOX, FOLDERS.SENT, FOLDERS.DRAFTS, FOLDERS.TRASH, FOLDERS.SPAM],
-		[]
-	);
 	const actionContainerRef = useRef<HTMLInputElement>(null);
+	const [open, setOpen] = useState(false);
+	const theme = useContext<ThemeContextProps>(ThemeContext);
 
-	const actions = useMemo(() => {
-		const arr = [];
+	const [visibleActionsCount, calculateVisibleActionsCount] = useVisibleActionsCount(
+		actionContainerRef,
+		{ numberLimit: maxActions }
+	);
 
-		if (message.parent === FOLDERS.DRAFTS) {
-			arr.push(sendDraft(message.id, message, t, dispatch));
-			arr.push(editDraft(message.id, folderId, t, replaceHistory));
-			arr.push(
-				moveMsgToTrash(
-					[message.id],
-					t,
-					dispatch,
-					createSnackbar,
-					deselectAll,
-					folderId,
-					replaceHistory,
-					message.conversation
-				)
-			);
-			arr.push(setMsgFlag([message.id], message.flagged, t, dispatch));
-		}
-		if (
-			message.parent === FOLDERS.INBOX ||
-			message.parent === FOLDERS.SENT ||
-			!includes(systemFolders, message.parent)
-		) {
-			// INBOX, SENT OR CREATED_FOLDER
-			arr.push(replyMsg(message.id, folderId, t, replaceHistory));
-			arr.push(replyAllMsg(message.id, folderId, t, replaceHistory));
-			arr.push(forwardMsg(message.id, folderId, t, replaceHistory));
-			arr.push(
-				moveMsgToTrash(
-					[message.id],
-					t,
-					dispatch,
-					createSnackbar,
-					deselectAll,
-					folderId,
-					replaceHistory,
-					message.conversation
-				)
-			);
-			arr.push(
-				setMsgRead([message.id], message.read, t, dispatch, folderId, replaceHistory, deselectAll)
-			);
-			arr.push(moveMessageToFolder([message.id], t, dispatch, false, createModal, deselectAll));
-			arr.push(printMsg(message.id, t, timezone));
-			arr.push(setMsgFlag([message.id], message.flagged, t, dispatch));
-			arr.push(redirectMsg(message.id, t, dispatch, createSnackbar, createModal, ContactInput));
-			arr.push(editAsNewMsg(message.id, folderId, t, replaceHistory));
-			arr.push(setMsgAsSpam([message.id], false, t, dispatch, replaceHistory));
-			arr.push(showOriginalMsg(message.id, t));
-		}
+	const firstActions = useMemo(
+		() =>
+			slice(actions, 0, visibleActionsCount > 0 ? visibleActionsCount - 1 : visibleActionsCount),
+		[actions, visibleActionsCount]
+	);
 
-		if (message.parent === FOLDERS.TRASH) {
-			arr.push(moveMessageToFolder([message.id], t, dispatch, true, createModal, deselectAll));
-			arr.push(deleteMessagePermanently([message.id], t, dispatch, createModal, deselectAll));
-		}
-		if (message.parent === FOLDERS.SPAM) {
-			arr.push(deleteMsg([message.id], t, dispatch, createSnackbar, createModal));
-			arr.push(setMsgAsSpam([message.id], true, t, dispatch, replaceHistory));
-			arr.push(printMsg(message.id, t, timezone));
-			arr.push(showOriginalMsg(message.id, t));
-		}
-		return arr;
-	}, [
-		message,
-		systemFolders,
-		t,
-		dispatch,
-		folderId,
-		replaceHistory,
-		createSnackbar,
-		deselectAll,
-		createModal,
-		timezone,
-		ContactInput
-	]);
+	const secondActions = useMemo(() => difference(actions, firstActions), [actions, firstActions]);
+	const iconSize = useMemo(() => parseInt(theme.sizes.icon.large, 10), [theme?.sizes?.icon?.large]);
 
-	const [visibleActionsCount, calculateVisibleActionsCount] = useOverflowCount(actionContainerRef);
+	const onIconClick = useCallback((ev: { stopPropagation: () => void }): void => {
+		ev.stopPropagation();
+		setOpen((o) => !o);
+	}, []);
+
+	const onDropdownClose = useCallback((): void => {
+		setOpen(false);
+	}, []);
+
+	const _maxWidth = useMemo(
+		() => (iconSize && maxActions ? `${iconSize * maxActions}px` : maxWidth),
+		[iconSize, maxActions, maxWidth]
+	);
 
 	useLayoutEffect(() => {
 		calculateVisibleActionsCount();
 	}, [calculateVisibleActionsCount]);
 
-	const firstActions = useMemo(
-		() => slice(actions, 0, visibleActionsCount - 1),
-		[actions, visibleActionsCount]
-	);
-	const secondActions = useMemo(() => difference(actions, firstActions), [actions, firstActions]);
-	const [open, setOpen] = useState(false);
-	const onIconClick = useCallback((ev: { stopPropagation: () => void }): void => {
-		ev.stopPropagation();
-		setOpen((o) => !o);
-	}, []);
-	const onDropdownClose = useCallback((): void => {
-		setOpen(false);
-	}, []);
 	return (
 		<Row
 			ref={actionContainerRef}
-			mainAlignment="flex-end"
-			takeAvailableSpace
+			mainAlignment={mainAlignment}
+			maxWidth={_maxWidth}
+			style={{ minWidth: '24px' }}
 			wrap="nowrap"
-			style={{ overflow: 'hidden' }}
+			takeAvailableSpace
 		>
-			{actions &&
+			{firstActions?.length > 0 &&
 				map(firstActions, (action) => (
-					<Tooltip key={`${message.id}-${action.icon}`} label={action.label}>
+					<Tooltip key={`${action.icon}`} label={action.label}>
 						<IconButton
 							size="small"
 							icon={action.icon}
@@ -224,12 +97,7 @@ const MailMsgPreviewActions: FC<MailMsgPreviewActionsType> = ({
 					</Tooltip>
 				))}
 			{secondActions?.length > 0 && (
-				<Dropdown
-					placement="right-end"
-					items={secondActions}
-					forceOpen={open}
-					onClose={onDropdownClose}
-				>
+				<Dropdown items={secondActions} forceOpen={open} onClose={onDropdownClose}>
 					<IconButton size="small" icon="MoreVertical" onClick={onIconClick} />
 				</Dropdown>
 			)}
